+++ src/index.js
@@ -10,7 +10,7 @@ import {
   Toast,
   ActionSheet,
   CheckboxGroup,
-  RadioGroup,
+  Radio,
   Slide,
   IndexList,
   Upload,
@@ -32,7 +32,7 @@ function install(Vue) {
     Tip,
     Toast,
     CheckboxGroup,
-    RadioGroup,
+    Radio,
     Slide,
     IndexList,
     ActionSheet,
