--- src/modules/radio/index.js
@@ -1,7 +1,0 @@
-import Radio from '../../components/radio/radio.vue'
-
-Radio.install = function (Vue) {
-  Vue.component(Radio.name, Radio)
-}
-
-export default Radio
